+++ legacy_one/trunk/src/hardware/hw_light.h
@@ -79,7 +79,7 @@ void HWR_DynamicShadowing(vxtx3d_t *clVerts, int nrClipVerts, player_t *p);
 void HWR_PlaneLighting(vxtx3d_t *clVerts, int nrClipVerts);
 void HWR_WallLighting(vxtx3d_t *wlVerts);
 void HWR_Reset_Lights(void);
-void HWR_Set_Lights(int viewnumber);
+void HWR_Set_Lights(byte viewnumber);
 
 #define DL_MAX_LIGHT    256  // maximum number of light (extra light are ignored)
